#70 refactor: create main repository class
